+++ backend/src/main/java/com/agrupae/application/exception/group/GroupNotClosedException.java
@@ -1,0 +1,9 @@
+package com.agrupae.application.exception.group;
+
+import com.agrupae.application.exception.ApplicationException;
+
+public class GroupNotClosedException extends ApplicationException {
+    public GroupNotClosedException() {
+        super("Group is not closed.");
+    }
+}